Fix tests for Auth0 module public interface
